+++ src/main/java/com/example/seo_dot/book/controller/MainPageController.java
@@ -2,7 +2,7 @@ package com.example.seo_dot.book.controller;
 
 import com.example.seo_dot.book.dto.response.BookListResponseDTO;
 import com.example.seo_dot.book.dto.response.NewBookListResponseDTO;
-import com.example.seo_dot.book.service.BookService;
+import com.example.seo_dot.book.service.MainPageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,17 +15,17 @@ import java.util.List;
 @RestController
 public class MainPageController {
 
-    private final BookService bookService;
+    private final MainPageService mainPageService;
 
     @GetMapping("/api/welcome/best")
     public ResponseEntity<List<BookListResponseDTO>> getBestSellers() {
-        List<BookListResponseDTO> response = bookService.getBestSeller();
+        List<BookListResponseDTO> response = mainPageService.getBestSeller();
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 
     @GetMapping("/api/welcome/new")
     public ResponseEntity<List<NewBookListResponseDTO>> getNewBooks() {
-        List<NewBookListResponseDTO> response = bookService.getNewBooks();
+        List<NewBookListResponseDTO> response = mainPageService.getNewBooks();
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 }
